:sparkle: consolidate db queries
